requirements.txt
streamlit
ollama
scikit-learn
pymupdf
python-docx
pandas
sentence-transformers
numpy
